DRM: await potential pending `generateRequest` and `load` call before closing a MediaKeySession
In the [EME specification](https://www.w3.org/TR/encrypted-media) it is
implicitely indicated that closing a `MediaKeySession` cannot be done
until `generateRequest` and `load` has been called on it.

Because we don't always trust platforms to perfectly implement the EME
specification (I would say that those are the APIs with which we
encounter the most issues), we had a less than ideal yet general solution
which just retried closing a MediaKeySession (with an exponential
backoff) when it couldn't be done due to an `InvalidStateError` - the
error rejected when the MediaKeySession is not yet ready to be closed.

However, there was multiple issues, especially on low-end devices, that
made me reconsider this solution:
  1. The exponential backoff could lead to large waiting time in some
     cases. This could in turn lead to longer content-loading time.
  2. If a `load` or `generateRequest` takes a LOT of time to respond
     (several seconds), we could be left in a case where we wouldn't be
     closing a session (because there's a limit to the amount of time
     we're re-trying to close a session) leading to a leak.
  3. Following the logic through logs in times of problems become very
     confusing for everyone involved in a debugging process

So this commit tries another solution instead.
What is done here is to always only call `MediaKeySession.close` once
(and fail if it fails) but ensuring that it doesn't happen while a
`generateRequest` or `load` call is pending. If one of those calls is
pending, we will actually call `close` when one of those two - which
both return Promises - either resolve or reject.

A `close` call before the first `generateRequest` or `load` call is
still authorized to maximize compatibility but once `close` has been
called, no `generateRequest` or `load` call will be performed.

This is enforced by always going through the `LoadedSessionStore` for
any of those calls, a solution that is not perfect, but does the job
easily.