DRM: Update persisted info after reloading a MediaKeySession
This commit allows to update information stored with a persisted
MediaKeySession if the format's version was anterior to what can be set
now.

As higher versions today always mean strictly more information, this
will allow to migrate re-loaded MediaKeySessions to the newer format
and thus, to progressively store more information on already-stored
session as they are re-loaded.