If seeking after the last potential position, load last segments before ending
This commit fixes a theoretical issue that could arise if an user seeked
after the last possible position.

Doing this is difficult to reproduce but is for example possible when
multiple audio or video AdaptationSets have different durations, with at
least one inferior to the corresponding Period and/or Manifest's
duration.
In that situation, the media element's duration may be set higher than
the maximum time on the shorter AdaptationSet and it will thus be
possible to seek further than it.

After the seek is done, the RxPlayer would notice that we are after the
last position, call the `endOfStream` API to notify that playback is
done, and switch the `ended` state.

Without putting much thoughts into it, this could seem like the right
thing to do. However, `endOfStream` will have the problematic
side-effect of updating the duration of the media element to the current
last buffered position (which would be the one before the seek was
done) and may show the corresponding last buffered frame at the time,
instead of the last frame of the content, which may be more logical.

---

To implement this, I chose for now to isolate the code of the
`RepresentationStream` where the time range of segments that should be
loaded is calculated, to include an exception when the position is after
the last Period if it is not still pending.
This exception would redirect the range so it loads starting from the
last second of the content.

I'm not too proud of this solution however as this kind of check is
unusual in the `RepresentationStream`. The main job of this module being
to just find which segments to load, to then load them and push them
without considering the grander scheme of how playback is going.

But it was so much easier to add this exception there, that I for now
decided to do that!