ABR: update PlaybackObserver's `liveGap` prop to just `maximumPosition`
This is a minor commit that just renames the `liveGap` as produced by
the PlaybackObserver to provide the distance to the live edge to a more
general `maximumPosition` which is just defined as the "theoretical
maximum position on the content that can currently be played.".

Both are actually completely equivalent under the current code, but
"live edge" implies that this was equal to the media data currently being
broadcasted or at least the last media position currently available which
would both be false assumptions (it is actually more: the maximum
position for which there's media data regardless of the choosen
Adaptation - yes it's a complex concept, I'm working on it now).

Because of the renaming, I had to add a `context` attribute to the
`ABRManager`'s code:
It was before able to tell if the current content was live by checking
the presence of the `liveGap` attribute, but with `maximumPosition` now
set regardless of if the content is live or not, the `ABRManager` would
now be unable to rely on this check.

So I added a way to communicate the `Manifest` structure to the
`ABRManager`.
I don't see this as a problem anyway as things become much more explicit
that way (and to me, that means better) and as it seems logical to me
that the `ABRManager` knows which `Manifest`, `Period` and `Adaptation`
it currently is choosing a `Representation` for.