Separate the concept of "livePosition" from the concept of "maximumPosition"
This commit continues our efforts to improve on the RxPlayer management
of the maximum position.
The idea here is to divide the previously-ambiguous `maximumPosition`
notion to two separate values:

  - `maximumSafePosition`: which is the maximum playable position
     regardless of the currently chosen Adaptation (track)

     It is basically the last position on which we're sure to have
     content in every audio and video tracks.

  - `livePosition`: which should be set to the edge of what is
    being broadcasted.
    Right now in DASH, it is just set to the maximum position that can be
    reached in the various audio and video Adaptations.

    It can be set to `undefined` for non-live contents.

This way instead of just relying on a vague `maximumPosition` to perform
very diverse operations such as calculating the distance to the live
edge, setting the content's duration, sending warnings when the current
position is over it and so on, we can instead use the more explicit value
of the two in function of what we want to do.