DASH: don't include presentationTimeOffset for the default initial time of a`<S>`
Some DASH MPD rely on `SegmentTimeline` elements to describe which
segments are available.

This `SegmentTimeline` contains `<S>` XML elements, indicating through
its attributes the time information (start time and duration) for either
a single segment or a group of consecutive segments.

Usually, at least the first of those `<S>` elements indicates a start
time, as a `t` attribute (this is rarely seen for the consecutive `<S>`
however, as here the starting time can simply be calculated by summing
the initial time and the other segment's duration).

However there could be cases where even the first `<S>` element does
not indicate a time. Here, it is to the RxPlayer to infer the time
associated to the first segment.

The previous operation considered both the Period's start time and the
optional `SegmentTemplate@presentationTimeOffset`, the latter being a
value signalling an offset between the time the segments were initially
for and the time it is intended to play at relatively to the Period's
start.

However, it seems (after comparing to packager output, to the
dash.js's behavior and doing some thinking!) that we should only rely to
the calculated Period's start to calculate the first initial time.

This commit fixes that.