Avoid infinite loop due to rouding errors while parsing multi-Periods MPDs
Fixes #1110.

This commit fixes an infinite loop that could occur on multi-Period MPDs
due to a rounding error linked to floating-point operations.

This possibily was very easy to work-around because one of the two
inner branches of that loop actually always led to a loop exit (because
it would always break the loop's condition - though here it didn't due
to that damned rounding error).

Explicity writing a `break` instruction when that branch is taken is
thus here a quick and easy fix, though it may not be the most elegant
way of dealing with it (if you have more simple and idiomatic ways, I'll
change to it).

I also added comments and inversed the inner conditions, in the hope I
made what that function is doing clearer.