DASH: Fix potential infinite rebuffering when a KID is not anounced in the MPD
We introduced a regression in the v3.27.0 which can lead to infinite
rebuffering/loading when playing MPD which do not anounce the key ids in
a `@cenc:default_KID attribute` (which seems to be required by the DASH-IF
IOP? That's why I understand from Chapter 7.6.2.1 of the v4.3, though
I'm unsure and we chose to be resilient anyway).

Basically the source of the error is a difference in semantics between
what our MPD parser parses and what the `ContentDecryptor` takes in
input to trigger the whole decryption logic.

Especially, the difference is how the `keyIds` property, defined in both
places is defined:

  - for the parser, it is the array of just the "key IDs associated with
    that content" and is always set as long as the content is encrypted.

  - for the `ContentDecryptor`, it's much more precize: it is the "key
    ids linked to [some] initialization data" - and more importantly:
    it is **"`undefined` when not known (different from an empty
    array - which would just mean that there's no key id involved)."**

So here, the parser is setting an empty array which means for it: "I
don't know".

Yet the `ContentDecryptor` receives that same information and it
understand: "There's no such concept of "key id" there".

---

To fix that situation, I chose what I think is the more compatible (less
error-prone) solution: to make it `undefined` by default and handle the
empty array as an ambiguous case in the `ContentDecryptor`.

I first added the possibility on the manifest parser-side
of defining an `undefined` `keyIds`, which would mean: "the key id(s)
associated with that content may exist but are not known" while the
`ContentDecryptor`'s definition doesn't change.

I also updated the MPD parser so it is always `undefined` if we didn't
find one in the MPD. This should prevent most `keyIds` array from
being empty (with the exception of offline "local" contents which may
still be).

Moreover, parts of the `ContentDecryptor` that made comparison of key
ids - mostly to check if a key id was already processes - now treat the
empty array as if its key ids are ambiguous: a Representation with an
empty `keyIds` array will never be assumed to already have been handled
for example, only that "we don't know if it has been handled or not just
based on its key id".

Thankfully, the `ContentDecryptor` is then generally resilient when it
cannot use key IDs and just rely on the whole initialization data
instead.