The demo page now use the Minimal RxPlayer (and optionally the WASM parser)
This commit is a proposal to use the Minimal RxPlayer (the one with
feature switching) instead of awkwardly modifying a `__FEATURES__`
global to chose which feature to include and using the "main" RxPlayer
exported.

As importing the minimal RxPlayer instead of the legacy default one is
something we advise to all applications, I feel it wasn't coherent to
not follow our own advice here!

In the end, I think we may want to get rid of the `__FEATURES__` object
altogether (we still need it now to define the wanted default feature
included in our legacy builds and for tests).

---

But mostly those modifications were just a [nice] side-effect of the
real improvement of this commit: I added to npm scripts,
`npm run start:wasm` and `npm run s:wasm` which are like respectively
`npm run start` and `npm run s` except they also import the `DASH_WASM`
feature (the WebAssembly-powered MPD parser), as I felt it was too
inconvenient to test until now.

This also means that I didn't make it the default. Why? For now I think
that checking the stability of the default JavaScript one is the
most important, though testing the WebAssembly one should also be
periodically done.

I wondered if we couldn't directly have a checkbox in the demo page
instead, but considering that you cannot "remove" the `DASH_WASM`
feature, doing this is in an understandable way is not easy.