Only log types which are inexpensive to stringify
In our ongoing efforts towards improving the "debuggability" of the
RxPlayer on devices with comparatively low performances, this commit
ensures that only values which are inexpensive to stringify and
lightweight are logged: short-enough strings, number, null, boolean,
undefined  (and Error instances as an exception).

This was done to limit the influence logging can have on performance,
especially preventing future logs from logging big objects or huge
`Uint8Array` like before.

As most debugging information we process is under a string form anyway, this
comes with only few disadvantages.

The main disadvantage I can think of is that we may have less precize
information when debugging through desktop/laptop's browsers' inspectors.
Though I thought that this may be improved by better logs and - for the
more precize data points - logs added manually on a case-by-case basis.