DASH: initial S elements without a `t` attribute implicitely begin at `0`
This is another attempt, after 161157a350de53debce09f3465834e3176d2b918,
to calculate a default `t` attribute for the first `<S>` element of a
DASH's `<SegmentTimeline>` when the first one omitted it (which happens
quite rarely).

The initial logic basically made as if the first segment had a time of
`Period@start - SegmentTemplate@presentationTimeOffset`. However, we did
have issues with it with some of those rare streams.

The 161157a350de53debce09f3465834e3176d2b918 fix tried to only consider
`Period@start` here to work-around those issues, this was mainly done
this way because it seemed to fix the encountered cases, as I found the
DASH specification unclear on that situation.

However, after doing one of our "beta-like" release (`dev` and `canal`
releases), we saw that this broke other streams - which had no problem
before.

This was when we actually tried to reverse engineer what other players
are doing.

It seems that the consensus is actually doing as if the first `t` is
actually set as if it was set to `0` (so we should actually go one step
further and consider neither inputs):
 - dash.js: https://github.com/Dash-Industry-Forum/dash.js/blob/2602917b9743cfae51fdd934f8e0beb9631a72a7/src/dash/utils/TimelineSegmentsGetter.js#L60
 - shaka-player: https://github.com/shaka-project/shaka-player/blob/4fc7a4893fd081d4dafa26a2034361afd7b7e6ed/lib/dash/mpd_utils.js#L142
   (the shaka-player already removes the presentationTimeOffset in that
   calculation, like we can see line 154 of the same file, so it ends up
   just being `-presentationTimeOffset` here).

In this commit, we thus chose to align with what those other players are
doing.