[Proposal] Extract key id from initialization segments if found
The RxPlayer has some advanced encryption-related features making use of
an encrypted Representation's key-id - which is an identifier for the
key used to decrypt the segment.

Such advanced features are for example the ability to obtain all
keys - excluding those for which the license server does not trust the
current device enough - a given content or DASH Period(s) needs from a
single license request and not just a single key. In turn, this allows
to reduce load and synchronization effects on the license server (as
well as potentially reducing the loading time of a content).

When key ids are not known (mostly when they aren't in the Manifest
file), the RxPlayer tries to stay resilient and should still be able to
provide most features, though things may not be as efficient as they could
be (e.g. multiple license requests could be performed instead of one).
Moreover, we already had bugs in the past linked to that exact situation
(e.g.: #1113).

Listing the exact influences the omission of key-id in the Manifest lead to is
difficult however, as layers of optimizations and resilience code made it hard
to clearly paint a picture of everything that would go on! But it probably has
multiple negative effects, even if the content can play.

---

Anyway, while working on another Proof-of-Concept, I noticed that the key id
of encrypted media was actually present in MP4 files, in the `tenc` box.

_In retrospect, this makes total sense, it may (or may not!) be how the
lower-level code/CDM is able to identify which key it should currently use
when multiple decryption keys are currently actively used (e.g. for
different qualities)._

---

This commit is thus a proposal to read the key id directly from the
initialization segment to obtain it even when it is not in the
Manifest file.

Moreover, if the key id is not in the Manifest, I disabled the
optimization which may lead to decryption negotiation being performed as
soon as the Manifest is loaded (comparatively to the default: after the
corresponding initialization segment is loaded) because the key id might
in most cases be known once the initialization segment is parsed, and that
in turn might lead to be able to deduce that a license request might not
even be needed (because the corresponding key has already been obtained).