DASH: set a minimum for the first available position of an index
The previous RxPlayer's behavior could lead to consider that the first
playable media data in a Period starts before a the Period's start, or
even before a `0` position which should not be possible.

In turn, this could lead to strange behaviors, like producing a negative
initial time (which seems to just lead to infinite rebuffering).

This commit ensure that an index never communicates it has media data
before the related Period's start.