Fix issues that could arise when handling segments starting at a negative position
After doing some experimentation with the `@presentationTimeOffset` attribute
of a DASH MPD to omit certain parts in the beginning of various
contents, we noticed that the RxPlayer had a poor handling of segments
starting at a negative time.

As media data before `0` will be ignored anyway, the easy solution is
just to make as if the media data before the `0` position did not exist.

This is what this commit does.