Do not re-download segments GCed by the browser
We introduced in #990 (v3.26.1) the notion of an history of pushed
segments to detect cases where the browser seems to garbage collect
parts of pushed segments.

The idea is simple (yet surprising, it might sound ugly at first!): we
register the recent history of segments pushed on browser-side buffers,
particularly their start time and end time as [indirectly] reported by
the browser [through the `buffered` TimeRanges object and some
deduction].
If we notice that at least two times in a row, the browser is garbage
collecting parts of a segment the same way (like 1 second at the
beginning for example), we stop trying to load that segment (as long as
the GCed pattern is the same one - if more/less is GCed, we may
retry loading it).

This works actually very well and fixed most segment requesting loop
occuring in the RxPlayer due to the browser acting weird - while still
allowing us to retain the current architecture and features of our
garbage-collection detection algorithm.

Yet it remains at least one credible case where it is still possible to
have requests loops of the same segment: when the browser is just
completely garbage collecting it each time (for the moment only seen
when doing very wrong things with segments, like removing any notion of
an intra video frame in one).

---

To fix this is however very straightforward: if previously we only added
an entry to the the buffered history when a segment's actual buffered start
and ends were known and set, we can now also add an entry if the segment was
not actually being considered as buffered after being pushed.

Then, when deciding which segments to load, we check for each segment the
possibility it was already pushed an immediately GCed. We only will skip
segments that have been GCed immediately two times in a row, because
sometimes browsers act up only temporarily (even sometimes for good reasons).

Though, as we're doing this with each segments at each playback
observation (for simplicity reasons), and as we didn't really try to work
on that code's performance yet, I'm wondering whether this has an
important-enough negative impact on perfs. I will test that now.