Reorganize core/api directory
This commit is a proposal for a small re-organization of the `src/core/api`
directory, according to the recent discussions we had on improving code
inspectability for newcomers:

  1. Both modules/files handling track switching (
     `media_element_track_choice_manager.ts` for "directfile"-based
     contents and `track_choice_manager.ts` for MediaSource-based contents)
     are moved to a `tracks_management` directory.

     The point is to de-clutter the `src/core/api` directory by grouping
     two very similar files together

  2. The small `get_player_state.ts` and `emit_seek_events.ts` files
     have been moved to a `utils.ts` file.

     The point here is to group those small utilitary functions into a
     file whose name make it clear that they just are small functions
     needed by the other modules.from two similar

For now, `option_utils` still has its own directory, I was unsure if I
should also move it to the `utils` one but thought its role was precize
enough and not that confusing.

Note that this proposal may only be a very small part of all modifications
we wish to do on all directories, depending on our discussions and if
these attempts seems promising.