Define all API-exposed types directly in src/public_types.ts
This commit is a proposal to fix at once the issue seen in #1076 where
various code editors propose "weird" path when auto-importing TypeScript
types.

Those paths lead actually to the RxPlayer's internal code (which is not
part of the API and might change at any version).

This is a problematic situation which seems to be frequent, we've seen
several applications at Canal+ also importing internal path that way -
incidentally (through auto-import) or not (e.g. the type is not
exported).

In turn this raises the bar to keep those applications build-able after
incrementing the RxPlayer's version, where under semver rules, they
should not have to care about that (it should just work),

---

The previous strategy to prevent that situation was to make sure that
most sensible type are exported and documented.
But developpers [these days ;p] prefer to just let their language server
do the import code by itself when using methods/events depending on that
type. Fair enough!

Making tsserver or whatever do auto-import of the right path is tricky.
It seems to be smart, sometimes too smart.

For example my first attempt was to declare proxy types in the exported
`src/public_types.ts` files, which then is relied on by every exposed
methods, events and properties.
It worked almost for all types, but in some situations, particularly
when the type was renamed when proxied, the language server actually
prefered the original, not renamed, type and went to that one instead
:/.

---

So this current work take a more drastic step:
All exposed types are here defined in the exported `src/public_types.ts`
(that's the file reachable trough the documented `"rx-player/types"`
path.

When any of the RxPlayer's code actually needs to use those types, they
will import it from that file.

As an advantage, it makes it clearer for us what types are actually part
of the API, preventing accidental breaking changes and helping us to
clearly see the complexity of it.

As a disadvantage, importing `public_types` deep inside the code for
types related to that actual code (e.g. decryption code importing
`public_types` to define the interface for loading persistent sessions
looks awkward).

Thoughts?