Update unit test files to be more idiomatic Jest
While trying to do the transition to Jest 28, I received warnings
linked to the mocking logic of our unit tests.

It was linked to the fact that closures implementing Module mocks used
objects declared outside of their scope, which is something we apparently
shouldn't do, unless the object was a mock whose name was prefixed by
the `mock` string.

I guess this is Jest trying to prevent a footgun when using dynamically
updated objects in mocks, as those might not be updated once already
imported (this is only a guess though).

Though, we already considered that possibility and had a common way to
name mocks, which was to add the suffix `Spy`, so this warning bothered
me a little!

Still, I understand that Jest wants to be opinionated and I thus did the
transition (from suffixing with `Spy` to prefixing by `mock`) so testing
code is more idiomatic with other Jest tests in the wild (even if I do
not like the idea of framework lock-in, especially for testing
frameworks - which mostly look like each others).

---

I also replaced most `require` calls in the unit testing code by
`jest.requireActual` function calls, to remove the awkward mix of ES6
and CommonJS imports we had before and make it more explicit why we
were doing that (which is to import the file *after* `jest.mock` module
mocks are applied on the code linked to that file - but not to that file
itself).
This looks even more than a lock-in that the other code update yet is
very easy to rollback.