Fix fallback for license missing kid in an extended singleLicensePer setting
We noticed of a regression that has been introduced in our DRM-related
code last release (a risk we took considering the complete DRM overhaul
occuring in it), where we could be left in an infinite rebuffering
situation under very specific conditions when a license is requested in
a complex `singleLicensePer` mode (`"periods"` or `"content"`).

For this bug to occur, we should be in a situation where we:

  1. have already loaded the same content recently or persisted its
     license in a complex `singleLicensePer` setting

  2. load that cached session when loading again that content, also
     under a complex `singleLicensePer` setting (note that we _might_
     also have had the issue in a simple `singleLicensePer`, still only
     if the first load was in a complex one)

  3. load as a first quality a content whose key-id was linked to the
     cached session but is not announced in the license (what we call
     in our jargon a "session's implicit key id").

     Typically this happens for the qualities for which the license
     server voluntarily retain the key(s) due to a lack of trust for the
     current device.

  4. never change quality through the adaptive logic (even choosing
     another quality to immediately come back fixes the issue).

In that condition, the key id missing from the cached session will not
be re-announced as missing, leading to the RxPlayer never going in its
fallbacking logic.

Note that this only happens for the absolute first quality chosen
because the issue is linked to the logic handling the initial announce of
what we call "implicit key ids".
It happens immediately after the license request, which is tightly linked
to the first encryption metadata encountered.

If the quality is chosen after, we go into another logic which checks if
that new quality might already be in a scope of an older session. That
logic works correctly.
Interestingly this means that switching qualities in the problematic
situation even only to come back, fixes the issue. This is because we go
through the working logic in that case.

---

To fix this issue, I changed the role of the `getFetchedLicenseKeysInfo`
function, whose role was mainly to find out implicit key ids linked to
sessions whose license are "fetched" (i.e. not sessions retrieved from
some cache) by using the `singleLicensePer` setting and the content the
license request was for, to a much more general
`getKeyIdsLinkedToSession` function whose role is the same thing but for
all kind of sessions (fetched AND loaded from any cache/storage).

That `getKeyIdsLinkedToSession` function also takes two supplementary
arguments:

  1. the `KeySessionRecord` of the session (key-id store linked to the
     MediaKeySession) - allowing it to know about possible
     previously-known implicit key ids associated to that session.

  2 and a `isCurrentLicense` boolean, that should be set to `true` in
    cases where the license was just fetched (opposed to loaded from
    cache) - allowing to know if the `singleLicensePer` scope applies there
    or if it's risky to use it.

It also checks (only in complex `singleLicensePer` modes due to fear of
loss of compatibility by adding this feature) if the key id(s) for which
the license request was performed is also in the license, even if the
license/session was just loaded from cache (that same logic only ran
for licences that had just been fetched previously). If it is not, it
returns it as an implicit key id.

Technically, this last bit should not be necessary (short explanation: if
the session was loaded from cache in the first place, the quality's key
id was probably in the session's `KeySessionRecord` which has already
been checked for implicit key ids at this point) but was still added as
a resiliance measure (we never know what future code updates could lead
to).