Don't switch to ENDED state if playback Observable finishes
As a resilience measure, it was decided that if the global playback
Observable completes, which it should never do, it should switch to
the ENDED state and clean up all its state.

This code should never be entered, because the Observable normally never
completes before we unsubcribe from it, so this was only meant as a
safety measure if that improbable case actually happened.

But since work done in f5867fc37e329eda5f973b275514694cbf5f5311,
the Observable does complete due to a side-effect where the `stoppedContent$`
Observable given to its `takeUntil` operator emits just before the
Observable is actually unsubcribed from.

This in turn led to the ENDED state being switched to before STOPPED every
time we wanted to stop the content.

To fix that problem I chose to:

  1. replace switching to the `ENDED` state to switching to the
     `STOPPED` state. It just makes more sense as the content has been
     completely stopped and cleaned up.

     Note that as this case should never happen, this does not change
     anything at the API-level

  2. We're now only doing that logic if the Observable completed
     **before** the RxPlayer actually stopped the corresponding content.

     This is very easy to check now that we use a `TaskCanceller` /
     `CancellationSignal` to explicitely stop the content.