Better handle reverse playback by not skipping very small holes when paused
An application using the RxPlayer to implement a frame-to-frame feature
where it's possible to go forward or backward, and also a fake reverse
playback, based on multiple very close seeks at a quick interval, is
having issues when seeking a very small amount of time backward at the
start boundary of a segment.

Once that boundary is reached, seeking back a very small amount of time
(let's say 0.04 second), actually lead to the RxPlayer seeking back
automatically to the start of the next segment in most cases.

This is because of an optimisation of the RxPlayer where:
  - if a very small amount of time in a segment is needed (less than 0.2
    seconds)
  - AND that segment is not yet loaded and buffered
  - AND if the next segment is already loaded and buffered

then we can directly seek to the start of the already buffered segment
instead, so playback can start right away (instead of having to wait the
loading of the previous segment).

This optimisation makes sense in a condition of regular playback but it
does break frame-to-frame use cases.

A simple and non-breaking tweak we could perform is to just perform that
kind of optimisation when the video is actually clearly playing in a
regular manner, that is it is not paused, at a `0` `playbackRate` or at
a negative `playbackRate`.

---

The same logic has been repeated for the "locked stream" handling, which
likewise seek at the start of the next Period in very specific
conditions.

Doing this here may be more problematic, as it might mean a visual
infinite rebuffering until either the player is unpaused or until the
application seeks again, though this is a much smaller inconvenience
that just not being able to seek back.

---

In our reflections, we also thought and even tried to implement
optimal reverse playback where e.g. media segments would be loaded in the
reverse order (a segment N, then a segment N-1 etc.), but it appeared that
doing that would not be trivial, for a usage that still stay limited.

So for now, this kind of more complete solution is put on hold, and
we'll make them test this [much much much] more simple but limited
solution (among issues, as segments are still loaded in regular order,
they might see a rebufferings after passing a segment boundary in the
reverse order, I'm also unsure of how it reacts with browser's garbage
collection algorithms which tend also to consider that playback goes in
the "regular direction" when choosing which segments to evict).